+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 * Only cross reference index enabled repositories.
 * Fixed an issue where search would return results with empty file contents for matches in submodules with indexing enabled. Searching over submodules is not supported yet, so these (empty) results have been removed.
 * Fixed an issue where match highlighting would be incorrect on lines that contained multibyte characters.
+* Fixed an issue where search suggestions would always link to master (and 404) even if the file only existed on a branch. Now suggestions always link to the revision that is being searched over.
 
 ## 2.5.16, 2.5.17